+++ src/index.ts
@@ -1,4 +1,4 @@
 export { Binder } from './binder';
 export { Container } from './container';
 export { Inject, Injectable, Optional, PostConstruct } from './decorators';
-export { ClassConstructor, InjectableId, Injector } from './injector';
+export { ClassConstructor, InjectableId, InjectionToken, Injector } from './injector';